pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "semantic-doc-indexer"
version = "1.0.0"
description = "A semantic document indexer using Ollama and ChromaDB"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=0.5.20",
    "ollama>=0.3.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "pypdf>=4.0.0",
    "pillow>=10.0.0",
    "python-docx>=1.0.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "mypy>=1.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
]

[project.scripts]
indexer = "src.indexer.main:run_indexer"

[tool.setuptools.packages.find]
where = ["src"]
include = ["indexer*"]

[tool.setuptools]
py-modules = []

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "chromadb.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pypdf.*"
ignore_missing_imports = true

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["src"]
skip = ["__init__.py"]